01_ai_agents_first/projects/DeepSearch/pyproject.toml
[project]
name = "news-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain-core>=0.3.74",
    "openai-agents>=0.2.4",
    "tavily-python>=0.7.10",
]
